Also no idea why this loses -370 elo

+++ src/typedefs.h
@@ -5,8 +5,8 @@
 
 typedef std::uint64_t u64;
 typedef std::uint32_t u32;
-typedef std::uint32_t u16;
-typedef std::uint32_t u8;
+typedef std::uint16_t u16;
+typedef std::uint8_t u8;
 typedef std::int16_t i16;
 
 #endif
